+++ src/main/java/com/photory/service/auth/AuthService.java
@@ -6,9 +6,11 @@ import com.photory.common.exception.model.NotFoundException;
 import com.photory.common.exception.model.ValidationException;
 import com.photory.common.util.JwtUtil;
 import com.photory.common.util.RedisUtil;
-import com.photory.controller.auth.dto.request.*;
+import com.photory.controller.auth.dto.request.AuthEmailCompleteRequestDto;
+import com.photory.controller.auth.dto.request.AuthEmailRequestDto;
+import com.photory.controller.auth.dto.request.SigninUserRequestDto;
+import com.photory.controller.auth.dto.request.ValidateEmailRequestDto;
 import com.photory.domain.user.User;
-import com.photory.domain.user.UserRole;
 import com.photory.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.mail.javamail.JavaMailSender;
@@ -75,24 +77,6 @@ public class AuthService {
         }
 
         redisUtil.setDataExpire(email, "1", 60 * 60 * 24L);
-    }
-
-    public void createUser(CreateUserRequestDto request) {
-        String email = request.getEmail();
-        String password = request.getPassword();
-        String nickname = request.getNickname();
-
-        if (userRepository.existsByEmail(email)) {
-            throw new ConflictException(String.format("이미 가입된 유저의 이메일 (%s) 입니다.", email), CONFLICT_USER_EXCEPTION);
-        }
-
-//       if (redisUtil.getData(email) != null && redisUtil.getData(email).compareTo("1") == 0) {
-        User user = User.of(email, passwordEncoder.encode(password), nickname, null, UserRole.ROLE_USER);
-
-        userRepository.save(user);
-//        } else {
-//            throw new UnAuthorizedException(String.format("인증이 완료되지 않은 이메일 (%s) 입니다.", email), UNAUTHORIZED_EMAIL_EXCEPTION);
-//        }
     }
 
     public String signinUser(SigninUserRequestDto request) {
